requirements.txt
asgiref==3.5.0
attrs==21.4.0
certifi==2021.10.8
dj-database-url==0.5.0
Django==3.2.13
django-appconf==1.0.5
django-discover-runner==1.0
django-elasticsearch-dsl==7.2.2
django-imagekit==4.1.0
django-rest-framework==0.1.0
django-role-permissions==3.1.1
djangorestframework==3.13.1
elasticsearch==7.17.3
elasticsearch-dsl==7.4.0
gunicorn==20.1.0
iniconfig==1.1.1
packaging==21.3
pilkit==2.0
Pillow==9.1.0
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.8
pytest==7.1.2
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.9
whitenoise==3.3.1
